+++ src/ImageResize/Png.php
@@ -12,11 +12,20 @@ namespace DBlackborough\GrabBag\ImageResize;
  */
 class Png extends AbstractResize
 {
-    protected $extension = '.png';
+    protected $extension;
 
     /**
-     * Set the required options for the image resizer. To allow batch processing we set the
-     * majority of the options in the constructor to allow reuse of the object
+     * Png constructor.
+     */
+    public function __construct()
+    {
+        $this->extension = '.png';
+
+        parent::__construct();
+    }
+
+    /**
+     * Set the required options for the image resizer.
      *
      * @param integer $width Required width for the new image
      * @param integer $height Required height for the new image
@@ -27,20 +36,21 @@ class Png extends AbstractResize
      * the image will be stretched to fit the desired canvas
      * @param array $canvas_color Canvas background color, passed in as an rgb array
      *
+     * @return AbstractResize
      * @throws \InvalidArgumentException If any of the params are invalid we throw an exception
      */
-    public function __construct(
+    public function setOptions(
         int $width,
         int $height,
         int $quality,
-        $maintain_aspect = false,
+        bool $maintain_aspect = true,
         array $canvas_color = array('r' => 255, 'g' => 255, 'b' => 255)
-    ) {
-        if ($quality < 1 || $quality > 9) {
-            throw new \InvalidArgumentException('Quality must be a value between 1 and 9');
+    ) : AbstractResize {
+        if ($quality < 0 || $quality > 9) {
+            throw new \InvalidArgumentException('Quality must be a value between 0 and 9');
         }
 
-        parent::__construct($width, $height, $quality, $maintain_aspect, $canvas_color);
+        parent::setOptions($width, $height, $quality, $maintain_aspect, $canvas_color);
     }
 
     /**